Add apply find rule command
