+++ src/main/java/seedu/algobase/commons/core/Messages.java
@@ -9,6 +9,7 @@ public class Messages {
     public static final String MESSAGE_INVALID_COMMAND_FORMAT = "Invalid command format! \n%1$s";
     public static final String MESSAGE_INVALID_PLAN_DISPLAYED_INDEX = "The Plan index provided is invalid";
     public static final String MESSAGE_INVALID_PROBLEM_DISPLAYED_INDEX = "The Problem index provided is invalid";
+    public static final String MESSAGE_INVALID_FIND_RULE_DISPLAYED_INDEX = "The Find Rule index provided is invalid";
     public static final String MESSAGE_INVALID_TASK_DISPLAYED_INDEX = "The Task index provided is invalid";
     public static final String MESSAGE_PROBLEMS_LISTED_OVERVIEW = "%1$d problems listed!";
     public static final String MESSAGE_INVALID_TAG_FORMAT = "Invalid tag format! \n";
